# api/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "0.4.0",
    "title" : "Trackless API",
    "contact" : {
      "name" : "wjtje",
      "url" : "https://www.wjtje.ga"
    },
    "license" : {
      "name" : "Apache 2.0 License",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths" : {
...
    "/user" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Get all the users from the system",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create a new user on the server",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "firstname" : {
                    "type" : "string"
                  },
                  "lastname" : {
                    "type" : "string"
                  },
                  "username" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  },
                  "groupId" : {
                    "type" : "integer"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created the new user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "userId" : {
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Something is wrong with your input"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/user/{userId}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Get infomation about a single user",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Something is wrong with your input"
          },
          "404" : {
            "description" : "The user is not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ],
        "description" : "You can use an '~' as a userId to list your own account."
      },
      "delete" : {
        "tags" : [ "user" ],
        "summary" : "Remove a single user",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Something is wrong with your input"
          },
          "404" : {
            "description" : "The user is not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ],
        "description" : "You can use an '~' as a userId to list your own account."
      },
      "patch" : {
        "tags" : [ "user" ],
        "summary" : "Change a users info",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "firstname" : {
                    "type" : "string"
                  },
                  "lastname" : {
                    "type" : "string"
                  },
                  "username" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Something is wrong with your input"
          },
          "404" : {
            "description" : "The user is not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ],
        "description" : "You can use an '~' as a userId to list your own account.\nYou don't need to edit every thing at once."
      }
    },
    "/work" : {
      "post" : {
        "tags" : [ "work" ],
        "summary" : "Save your work on the server",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "locationId" : {
                    "type" : "integer"
                  },
                  "date" : {
                    "type" : "string"
                  },
                  "time" : {
                    "type" : "integer"
                  },
                  "description" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "workId" : {
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "your input values are wrong"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
    "/work/user/{userId}" : {
      "get" : {
        "tags" : [ "work" ],
        "summary" : "Get all the work from a user",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Work"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Check your input"
          },
          "404" : {
            "description" : "User not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    },
...
    "/work/user/{userId}/{workId}" : {
      "get" : {
        "tags" : [ "work" ],
        "summary" : "Get a single work object",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "workId",
          "description" : "Id of the wokr",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Work"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Check your input"
          },
          "404" : {
            "description" : "User or work not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "work" ],
        "summary" : "Remove a single work object",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "workId",
          "description" : "Id of the wokr",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Check your input"
          },
          "404" : {
            "description" : "User or work not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "work" ],
        "summary" : "Edit a single work object",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "description" : "ID of the user",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "workId",
          "description" : "Id of the wokr",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "locationId" : {
                    "type" : "integer"
                  },
                  "date" : {
                    "type" : "string"
                  },
                  "time" : {
                    "type" : "integer"
                  },
                  "description" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Check your input"
          },
          "404" : {
            "description" : "User or work not found"
          }
        },
        "security" : [ {
          "bearer_token" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Location" : {
        "type" : "object",
        "properties" : {
          "locationId" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "place" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          }
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "integer"
          },
          "groupName" : {
            "type" : "string"
          }
        }
      },
      "Work" : {
        "type" : "object",
        "properties" : {
          "workId" : {
            "type" : "integer"
          },
          "time" : {
            "type" : "integer"
          },
          "date" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/User"
          },
          "location" : {
            "$ref" : "#/components/schemas/Location"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearer_token" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  },
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/wjtje/Trackless/0.4.0"
  } ]
}